Update and rename LICENSE to LICENSE.md
This commit adds the CC BY 4.0 license to the repository. It documents that the content within is based on Abdul Kuyateh’s personal learning experience on TryHackMe. The license allows others to share and adapt the material with proper attribution. This ensures transparency, supports open learning, and clarifies that the content is not affiliated with TryHackMe.